refactor: Refeicoes para Refeicao
