+++ FoodMinder/src/main/java/com/foodminder/FoodMinder/domain/refeicao/RefeicaoRepository.java
@@ -1,5 +1,11 @@
-package com.foodminder.FoodMinder.domain.refeicoes;
+package com.foodminder.FoodMinder.domain.refeicao;
 
+import com.foodminder.FoodMinder.domain.planejamento.Planejamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RefeicoesRepository extends JpaRepository<Refeicoes, Integer> {}
+import java.util.List;
+
+public interface RefeicaoRepository extends JpaRepository<Refeicao, Integer> {
+    //List<Refeicao> findById(Integer id);
+
+}
