+++ FoodMinder/src/main/java/com/foodminder/FoodMinder/domain/refeicao/RequestRefeicao.java
@@ -1,18 +1,13 @@
-package com.foodminder.FoodMinder.domain.refeicoes;
+package com.foodminder.FoodMinder.domain.refeicao;
 
-import com.foodminder.FoodMinder.domain.TipoRefeicao.TipoRefeicao;
 import com.foodminder.FoodMinder.domain.planejamento.Planejamento;
 import jakarta.validation.constraints.NotNull;
 
-import java.util.List;
-
-public record RequestRefeicoes(
+public record RequestRefeicao(
 		Integer id,
 		@NotNull
 		String nome,
 		@NotNull
-		String receita,
-		@NotNull
-		Planejamento planejamento // Adicionando o atributo para representar o tipo de refeição
+		String receita
 ) {
 }
